testing colors

+++ ribbitx/cli.py
@@ -39,23 +39,22 @@ def ping():
     ping_response = requests.get(f"{base_uri}/system/ping", headers=headers)
     status_code = ping_response.status_code
     http_status = str(ping_response.content)
-    click.echo(f"HTTP Response {status_code}")
+    click.echo(click.style(f"HTTP Response {status_code}", fg="green"))
     if status_code == 200:
-        click.echo("HTTP Status OK")
+        click.echo(click.style("HTTP Status OK", fg="green"))
     else:
-        click.echo("Please check instance. Potentially Unhealthy.")
+        click.echo(click.style("Please check instance. Potentially Unhealthy.", fg="red"))
 
 
 # version command to check version of hosted instance
 @click.command("version", short_help="Check version of Artifactory instance.")
 def version():
     """Checks version of Artifactory instance."""
-    click.echo("Getting remote instance version...")
+    click.echo(click.style("Getting remote instance version...", fg="green"))
     version_response = requests.get(f"{base_uri}/system/version", headers=headers)
     version_json = version_response.json()
     click.echo("Artifactory Version: " + version_json["version"])
     click.echo("Artifactory Revision: " + version_json["revision"])
-    click.echo(version_response.content)
 
 
 # storage command to check status of hosted instance storage
@@ -93,7 +92,9 @@ def repocreate():
     repocreate_status = repocreate_request.status_code
     repocreate_json = repocreate_request.json
     if repocreate_status == 200:
-          click.echo(repocreate_request.content)
+      click.echo(click.style(f"Successfully created repository '{repo_name}'", fg="green"))
+    elif repocreate_status == 409:
+      click.echo(click.style("409 Error: An invalid character is being used.", fg="bright_red"))
     else:
       click.echo(repocreate_request.content)
 
@@ -102,14 +103,14 @@ def repocreate():
 @click.command("repo-update", short_help="Update a repository's public description.")
 def repoupdate():
     """Update a repository description."""
-    click.echo("This tool will allow you to update the public description of a selected local repository.")
+    click.echo(click.style("This tool will allow you to update the public description of a selected local repository.", fg="green"))
     repo_name = input("Please enter the repository name you are updating: ")
     repo_description = input("Enter the new description: ")
     repo = { "key": repo_name, "description": repo_description }
     repoupdate_request = requests.post(f"{base_uri}/repositories/" + repo["key"], json=repo, headers=headers)
     repoupdate_status = repoupdate_request.status_code
     if repoupdate_status == 200:
-      click.echo("The repo public description has been updated!")
+      click.echo(click.style("The repo public description has been updated!", fg="green"))
     else:
       click.echo("There was an error. Repository not updated.")
 
@@ -125,27 +126,34 @@ def usercreate():
   new_user_request = requests.put(f"{base_uri}/security/users/" + user_info["name"], json=user_info, headers=headers)
   new_user_status = new_user_request.status_code
   if new_user_status == 201:
-    click.echo(f"The user: {new_username} has been created.")
+    click.echo(click.style(f"The user: {new_username} has been created.", fg="green"))
   else:
-    click.echo("There has been an error. User already exists or try a stronger password.")
+    click.echo(click.style("There has been an error. Please make sure email address is correct or try a stronger password.", fg="bright_red"))
 
 
 # delete user command
 @click.command("user-delete", short_help="Delete a user.")
 def userdelete():
     """Delete a user."""
-    username = input("What is the username: ")
-    deleteuser_request = requests.delete(f"{base_uri}/security/users/{username}", headers=headers)
-    deleteuser_status = (deleteuser_request.status_code)
-    if deleteuser_status == 200:
-      click.echo(f"{username} deleted!")
-      click.echo(deleteuser_request.content)
-    elif deleteuser_status == 401:
-      click.echo(f"You are not authorized to delete user {username}")
-    elif deleteuser_status == 404:
-      click.echo(f"User {username} not found.")
+    username = input("What is the username you wish to delete: ")
+    click.echo(click.style(f"Are you sure you want to delete {username}'s account? [y/n] ", fg="yellow"))
+    delete_user = click.getchar()
+    click.echo()
+    if delete_user == "y":
+      deleteuser_request = requests.delete(f"{base_uri}/security/users/{username}", headers=headers)
+      deleteuser_status = (deleteuser_request.status_code)
+    elif delete_user == "n":
+      click.echo("Deletion Canceled")
     else:
-      click.echo("User not deleted.")
+      click.echo("Invalid Input. Try again.")
+    if deleteuser_status == 200:
+      click.echo(click.style(f"The user: {username} has been removed successfully.", fg="green"))
+    elif deleteuser_status == 401:
+      click.echo(click.style(f"You are not authorized to delete user {username}", fg="bright_red"))
+    elif deleteuser_status == 404:
+      click.echo(click.style(f"Error: User {username} not found.", fg="bright_red"))
+    else:
+      click.echo(click.style("Error: User not deleted.", fg="bright_red"))
